refactor types

+++ example/type.go
@@ -20,12 +20,12 @@ type TypeList struct {
 	FixedLenByteArray string  `parquet:"name=FixedLenByteArray, type=FIXED_LEN_BYTE_ARRAY, length=10"`
 
 	Utf8            string `parquet:"name=utf8, type=UTF8, encoding=PLAIN_DICTIONARY"`
-	Int_8           int32  `parquet:"name=int_8, type=INT_8"`
-	Int_16          int32  `parquet:"name=int_16, type=INT_16"`
+	Int_8           int8  `parquet:"name=int_8, type=INT_8"`
+	Int_16          int16  `parquet:"name=int_16, type=INT_16"`
 	Int_32          int32  `parquet:"name=int_32, type=INT_32"`
 	Int_64          int64  `parquet:"name=int_64, type=INT_64"`
-	Uint_8          uint32 `parquet:"name=uint_8, type=UINT_8"`
-	Uint_16         uint32 `parquet:"name=uint_16, type=UINT_16"`
+	Uint_8          uint8 `parquet:"name=uint_8, type=UINT_8"`
+	Uint_16         uint16 `parquet:"name=uint_16, type=UINT_16"`
 	Uint_32         uint32 `parquet:"name=uint_32, type=UINT_32"`
 	Uint_64         uint64 `parquet:"name=uint_64, type=UINT_64"`
 	Date            int32  `parquet:"name=date, type=DATE"`
@@ -71,12 +71,12 @@ func main() {
 			FixedLenByteArray: "HelloWorld",
 
 			Utf8:            "utf8",
-			Int_8:           int32(i),
-			Int_16:          int32(i),
+			Int_8:           int8(i),
+			Int_16:          int16(i),
 			Int_32:          int32(i),
 			Int_64:          int64(i),
-			Uint_8:          uint32(i),
-			Uint_16:         uint32(i),
+			Uint_8:          uint8(i),
+			Uint_16:         uint16(i),
 			Uint_32:         uint32(i),
 			Uint_64:         uint64(i),
 			Date:            int32(i),
